game-api/agents/npc_agent.py
import torch
# User mentioned PyTorch, so we'll prepare for it here.

class NPCAgent:
    def __init__(self, name: str, personality: str):
        self.name = name
        self.personality = personality
        self.memory = []

    def generate_response(self, user_input: str) -> str:
        # Placeholder for AI logic
        return f"{self.name}: I hear you saying '{user_input}'. My personality is {self.personality}."
