game-api/requirements.txt
fastapi[standard]>=0.115.8
uvicorn>=0.34.0
websockets>=14.0
langchain-core>=0.3.34
langchain-groq>=0.2.4
pydantic>=2.10.6
pydantic-settings>=2.7.1
python-dotenv>=1.0.1
loguru>=0.7.3